KotlinSprint-2-3 task new

+++ src/main/kotlin/lesson_2/less2_task1.kt
@@ -2,12 +2,16 @@ package org.example.lesson_2
 
 fun main(){
     val allStudent = 4f
-    val student1Assessment = 3f
-    val student2Assessment = 4f
-    val student3Assessment = 3f
-    val student4Assessment = 5f
+    val student1Assessment = 3
+    val student2Assessment = 4
+    val student3Assessment = 3
+    val student4Assessment = 5
 
-    val middleAssessment: Float = (student1Assessment + student2Assessment + student3Assessment + student4Assessment)/allStudent
+    val middleAssessment: Float = (
+            student1Assessment.toFloat() +
+                    student2Assessment.toFloat() +
+                    student3Assessment.toFloat() +
+                    student4Assessment.toFloat())/allStudent
 
     print(middleAssessment)
